+++ simplesat/sat/policy/__init__.py
@@ -4,5 +4,9 @@ from .policy import DefaultPolicy
 from .undetermined_clause_policy import (
     LoggedUndeterminedClausePolicy, UndeterminedClausePolicy
 )
+from .priority_queue_policy import (
+    LoggedPriorityQueuePolicty, PriorityQueuePolicy
+)
 
-InstalledFirstPolicy = LoggedUndeterminedClausePolicy
+#  InstalledFirstPolicy = LoggedUndeterminedClausePolicy
+InstalledFirstPolicy = LoggedPriorityQueuePolicty