+++ CHANGES.rst
@@ -10,6 +10,9 @@ Enhancements
 
 * Details relating to unsatisfiable scenarios are captured in an ``UNSAT``
   object and attached to the ``SatisifiabilityError`` raised (#101).
+* Adds a second policy based on a priority queue (#131)
+* Major speed improvements in the policies and assignment tracker (#131)
+
 
 Bugs Fixed
 ----------
@@ -17,6 +20,7 @@ Bugs Fixed
 * ``IPolicy`` constructor now ignores initialization arguments (#101).
 * Some sort operations that were using non-unique keys have been fixed (#101).
 * Assumptions are now represented as an empty Clause object (#101).
+
 
 Version 0.1.0
 =============